requirements.txt
fastapi >= 0.109.2
requests >= 2.31.0
uvicorn >= 0.27.1, <1.0.0
pandas >= 2.2.0
scikit-learn >= 1.4.1.post1
google-api-python-client >= 2.123.0
google-auth-httplib2 >= 0.2.0
google-auth-oauthlib >= 1.2.0
